+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Updated
+
+- Validation for TRIM to optionally require `location` and `actionDefinition`
+
 ## [1.0.1] - 2022-12-12
 
 ### Added